Separated monitoring to two scripts, grouped common fns to same file, added reporting function
